Bootstrap jenkins from scratch -- the oy vey! edition
This functionality had bitrotted because we've never
really re-installed jenkins from scratch

With this PR, I'm able to bootstrap and fully configure
Jenkins ver. 1.611 on a ubuntu utopic vagrant box.

The provision has three phases: init, config, jenkins.

The config phase is needed when bootstrapping on EC2,
where chef vault is not yet available. After the instance
is up and authorized with chef-server for vault access,
we can go to the config phase.

Once Jenkins has the github oauth plugin and is restarted, we
can run the master-jenkins recipe.

NOTES:
 - switch to chef_vault_item for development in vagrant
 - Pass `hudson.model.User.allowNonExistentUserToLogin` to JVM
   to open jenkins up a bit for provisioning
 - Also need `org.eclipse.jetty.server.Request.maxFormContentSize=1000000`
   to allow plugin installs (??)