use a newer version of Git (2.5.3, not 1.x)
I'm not sure this is absolutely necessary, but the version we were
using was old, and when I tested my longpaths fix, I was testing it
using Git 2.5.3, and I'd rather not have to redo the testing with
the older Git version.

putting the version number in the path, instead of just using "Git",
documents which version of Git we're using.

how it was installed: I ran the GUI installer (through CoRD)