use newer sbt-extras
the one we've been using is long in the tooth.

my motivation for this is, I would like to be able to use sbt to
write scripts, instead of everything being bash.  but my job
is hanging when it tries to run a `#!/usr/bin/env sbt` script
and I'm hoping this will fix it